Parametrize wikidata.org and WDQS
This is the first step of making it possible to run
this code on other, non-Wikidata-based sites (e.g. OSM wiki or Commons)

A follow-up is still needed to have a global site configuration,
preferably passable as a parameter to all classes that need it.

Note that in most cases the values are defined on instances, not static.
This should help with future per-instance configuration and testing.